refactor(react-wrapper): address review comments for settings mapper
