+++ wrappers/react-wrapper/src/helpers.tsx
@@ -212,6 +212,74 @@ export function isCSR(): boolean {
   return typeof window !== 'undefined';
 }
 
+function isObjectLike(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null;
+}
+
+function isArray(value: unknown): value is unknown[] {
+  return Array.isArray(value);
+}
+
+/**
+ * Deep-compare plain objects and arrays used by wrapper settings.
+ *
+ * @param {unknown} previousValue Previous value.
+ * @param {unknown} currentValue Current value.
+ * @returns {boolean} `true` if values are equivalent.
+ */
+export function areEquivalentSettingsValue(previousValue: unknown, currentValue: unknown): boolean {
+  if (previousValue === currentValue) {
+    return true;
+  }
+
+  const previousIsArray = isArray(previousValue);
+  const currentIsArray = isArray(currentValue);
+
+  if (previousIsArray && currentIsArray) {
+    const previousEntries = previousValue as unknown as unknown[];
+    const currentEntries = currentValue as unknown as unknown[];
+
+    if (previousEntries.length !== currentEntries.length) {
+      return false;
+    }
+
+    for (let index = 0; index < previousEntries.length; index++) {
+      if (!areEquivalentSettingsValue(previousEntries[index], currentEntries[index])) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  if (previousIsArray || currentIsArray) {
+    return false;
+  }
+
+  if (!isObjectLike(previousValue) || !isObjectLike(currentValue)) {
+    return false;
+  }
+
+  const previousKeys = Object.keys(previousValue);
+  const currentKeys = Object.keys(currentValue);
+
+  if (previousKeys.length !== currentKeys.length) {
+    return false;
+  }
+
+  for (const key of previousKeys) {
+    if (!Object.prototype.hasOwnProperty.call(currentValue, key)) {
+      return false;
+    }
+
+    if (!areEquivalentSettingsValue(previousValue[key], currentValue[key])) {
+      return false;
+    }
+  }
+
+  return true;
+}
+
 /**
  * A variant of useEffect hook that does not trigger on initial mount, only updates
  *
